restructure and begin working on docs
