+++ README.rst
@@ -1,0 +1,9 @@
+Template for the Read the Docs tutorial
+=======================================
+
+This GitHub template includes fictional Python library
+with some basic Sphinx docs.
+
+Read the tutorial here:
+
+https://docs.readthedocs.io/en/stable/tutorial/
